requirements.txt
flask
pywebview
pillow
pyinstaller
